Added "Aim Game" project
So this started out as me messing around with JavaScript but the idea evolved into something like an aim game.
Whenever you click the button it changes position randomly (Haven't tested this on displays bigger than 1080p!)
There's a score counter in the top left which keeps count of the times you've clicked the button. The idea is that when you first click the button a countdown between 15-30 seconds starts and during that time each one of your clicks will be counted and at the end of the countdown your accuracy % will be shown.